// types.ts
export interface Card {
  rowNumber: number;
  id: string;
  deckName: string;
  sound: string;
  image: string;
  noteEng: string;
  noteRus: string;
  explanationEng: string;
  explanationRus: string;
  japanesePhrase: string;
  japaneseSentence: string;
  romaji: string;
  japaneseSentenceFurigana: string;
  explanationRusDetailed: string;
  externalLinkRus: string;
}